NoEscape turned off by default
It is agitating if one loses a game on FICS simply because he disconnected (lost connection...etc). Setting noescape to off will try to avoid this problem and the user will not lose the game after losing connection but instead..the game will be adjourned and the user may either resign, abort, offer a draw or continue the game after this, leading to a more comfortable experience on PyChess. 

No functional change.